+++ openpype/hosts/blender/api/pipeline.py
@@ -1,6 +1,8 @@
 import os
 import sys
 import traceback
+from pathlib import Path
+from datetime import datetime
 from typing import Callable, Dict, Iterator, List, Optional
 
 import bpy
@@ -11,8 +13,9 @@ from . import ops
 import pyblish.api
 
 from openpype.client import get_asset_by_name
-from openpype.hosts.blender.utility_scripts.is_workfile_out_of_date import (
-    is_work_file_out_of_date,
+from openpype.client.entities import (
+    get_last_version_by_subset_name,
+    get_asset_by_name,
 )
 from openpype.pipeline import (
     schema,
@@ -30,9 +33,10 @@ from openpype.lib import (
     register_event_callback,
     emit_event
 )
+from openpype.lib.dateutils import get_timestamp
 import openpype.hosts.blender
 from openpype.settings import get_project_settings
-
+from .workio import current_file
 
 HOST_DIR = os.path.dirname(os.path.abspath(openpype.hosts.blender.__file__))
 PLUGINS_DIR = os.path.join(HOST_DIR, "plugins")
@@ -206,7 +210,7 @@ def on_open():
     if set_frames_startup:
         set_frame_range(data)
 
-    if is_work_file_out_of_date():
+    if not is_work_file_out_of_date():
         bpy.context.window_manager.is_workfile_out_of_date = True
         bpy.ops.wm.workfile_out_of_date("INVOKE_DEFAULT")
 
@@ -309,6 +313,34 @@ def _discover_gui() -> Optional[Callable]:
             return gui
 
     return None
+
+
+def is_work_file_out_of_date() -> bool:
+    """Check if the current workfile is out of date.
+    This is based on last modification date, so if a user modifies an out of
+    date workfile, this will return `False`. Also, in case of partial publish,
+    this will return `True`.
+
+    Returns:
+        bool: True if the current workfile is out of date.
+    """
+
+    session = legacy_io.Session
+
+    # Getting date and time of the latest published workfile
+    last_published_time = get_last_version_by_subset_name(
+        legacy_io.active_project(),
+        f"workfile{session.get('AVALON_TASK')}",
+        asset_name=session.get("AVALON_ASSET"),
+    )["data"]["time"]
+
+    # Getting date and time of the latest locally installed workfile
+    # Time is converted to use the same format as for `last_publishd_time`
+    workfile_time = get_timestamp(
+        datetime.fromtimestamp(Path(current_file()).stat().st_mtime)
+    )
+
+    return last_published_time > workfile_time    
 
 
 def add_to_avalon_container(container: bpy.types.Collection):
